rename -> phys_nonintersection
