+++ src/phys_nonintersection.ts
