+++ src/smoothing.ts
@@ -1,6 +1,6 @@
 import { EntityManager, EM, Component } from "./entity-manager.js";
 import { vec3, quat, mat4 } from "./gl-matrix.js";
-import { WorldFrameDef } from "./phys_esc.js";
+import { WorldFrameDef } from "./phys_nonintersection.js";
 import { tempQuat, tempVec } from "./temp-pool.js";
 import { Timer, PhysicsTimerDef } from "./time.js";
 import {
